requirements.txt
discord.py
python-dotenv
aiohttp
requests
mistralai


#pip install -r requirements.txt
